backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
sqlalchemy>=1.4.23
psycopg2-binary>=2.9.1
pydantic>=1.8.2
python-dotenv>=0.19.0
celery>=5.2.3
redis>=4.1.0
httpx>=0.23.0
python-multipart>=0.0.5
aiofiles>=0.8.0
openai>=1.0.0
anthropic>=0.3.0
google-generativeai>=0.2.0
groq>=0.3.0
pillow>=9.0.0
opencv-python>=4.5.5
numpy>=1.21.0
black>=22.3.0
pylint>=2.12.2
langchain>=0.1.0 